First part of passionate reservations
It was necessary to create this branch to realign the "veterinarian" branch, previously used to modify the code that allows the insertion of a new reservation by the veterinarian, to a more recent version of the code, which also sees changes in the "Passionate" part of the app. The reallign includes among other things a new type of dialog for uploading diagnoses for appointments that have already taken place.
In the next commit, all the work done since the last commit on the "veterinarian" branch will be made available here, to then be merged into the "master" branch and continue implementing other app features.

Co-Authored-By: Giacomo Detomaso <57029496+GiacomoDetomaso@users.noreply.github.com>
Co-Authored-By: Roberto <56017760+UnderScorriX@users.noreply.github.com>